+++ lib/database/index.d.ts
@@ -10,5 +10,6 @@ export declare class NanoSQLDefaultBackend implements NanoSQLPlugin {
     doExec(execArgs: IdbQuery, next: (execArgs: IdbQuery) => void): void;
     dumpTables(tables?: string[]): any;
     importTables(tables: any): any;
+    willDisconnect(next: any): void;
     extend(next: (args: any[], result: any[]) => void, args: any[], result: any[]): void;
 }