+++ src/database/adapter-indexedDB.ts
@@ -1,7 +1,7 @@
 import { NanoSQLStorageAdapter, DBKey, DBRow, _NanoSQLStorage } from "./storage";
 import { DataModel } from "../index";
 import { setFast } from "lie-ts";
-import { StdObject, hash, fastALL, deepFreeze, uuid, timeid, _assign, generateID, sortedInsert } from "../utilities";
+import { StdObject, hash, fastALL, deepFreeze, uuid, timeid, _assign, generateID, sortedInsert, intersect } from "../utilities";
 import { DatabaseIndex } from "./db-idx";
 
 const _evalContext = (source: string, context: {[key: string]: any}) => {
@@ -124,11 +124,11 @@ export class _IndexedDBStore implements NanoSQLStorageAdapter {
         this._dbIndex[tableName] = new DatabaseIndex();
 
         dataModels.forEach((d) => {
-            if (d.props && d.props.indexOf("pk") > -1) {
+            if (d.props && intersect(["pk", "pk()"], d.props)) {
                 this._pkType[tableName] = d.type;
                 this._pkKey[tableName] = d.key;
 
-                if (d.props && d.props.indexOf("ai") > -1 && (d.type === "int" || d.type === "number")) {
+                if (d.props && intersect(["ai", "ai()"], d.props) && (d.type === "int" || d.type === "number")) {
                     this._dbIndex[tableName].doAI = true;
                 }
             }