+++ src/database/adapter-levelDB.ts
@@ -1,7 +1,7 @@
 import { NanoSQLStorageAdapter, DBKey, DBRow, _NanoSQLStorage } from "./storage";
 import { DataModel } from "../index";
 import { setFast } from "lie-ts";
-import { StdObject, hash, fastALL, deepFreeze, uuid, timeid, _assign, generateID, sortedInsert } from "../utilities";
+import { StdObject, hash, fastALL, deepFreeze, uuid, timeid, _assign, generateID, sortedInsert, intersect } from "../utilities";
 import { DatabaseIndex } from "./db-idx";
 
 declare var global: any;
@@ -108,13 +108,13 @@ export class _LevelStore implements NanoSQLStorageAdapter {
         });
 
         dataModels.forEach((d) => {
-            if (d.props && d.props.indexOf("pk") > -1) {
+            if (d.props && intersect(["pk", "pk()"], d.props)) {
                 this._pkType[tableName] = d.type;
                 this._pkKey[tableName] = d.key;
                 this._isPKnum[tableName] = ["int", "number", "float"].indexOf(d.type) !== -1;
             }
 
-            if (d.props && d.props.indexOf("ai") > -1 && d.props.indexOf("pk") > -1 && d.type === "int") {
+            if (d.props && intersect(["ai", "ai()"], d.props) && intersect(["pk", "pk()"], d.props) && d.type === "int") {
                 this._dbIndex[tableName].doAI = true;
             }
         });